+++ CHANGELOG.md
@@ -50,6 +50,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - New updateBatch method to webhooks.subscriptionsApi
   - updated timelineModels
 
+### Updated
+
+  - extended webhooks-sample-app to setup application webhooks configuration after successful authorization
+  
+### Fixed
+
+  - 'apiKey' authorization for webhooks clients  
+
 [unreleased]: https://github.com/HubSpot/hubspot-api-nodejs/compare/v1.0.0-beta...HEAD
 [1.0.0-beta]: https://github.com/HubSpot/hubspot-api-nodejs/releases/tag/v1.0.0-beta
 [1.1.0-beta]: https://github.com/HubSpot/hubspot-api-nodejs/releases/tag/v1.1.0-beta